refactor : Level Domain refactoring

+++ backend/dduishu/src/main/java/com/b208/dduishu/domain/character/entity/CharacterLevel.java
@@ -26,7 +26,7 @@ public class CharacterLevel {
         this.exp = exp;
     }
 
-    public void addExp(float exp ) {
+    public void addExp(double exp ) {
         this.exp += exp;
         checkLevel();
     }
